requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
python-dotenv
langchain-groq
